fix: move enums and interfaces to types.ts
