+++ client/src/components/pages/lobby-page/game-settings-section/game-settings-section.tsx
@@ -1,39 +1,9 @@
 import { ChangeEvent, useEffect, useState } from 'react';
 import { makeStyles, MenuItem, Select, Switch, TextField } from '@material-ui/core';
 import SecondsTimePicker from '@components/ui/time-picker/time-picker';
+import { CardSet, IGameSettingsErrors } from '@models/types';
+import { cardSetData } from './cardSetData';
 import './game-settings-section.sass';
-
-export interface IGameSettingsErrors {
-  /*
-   * type of score error
-   */
-  scoreTypeError?: boolean;
-  /*
-   * type of score error
-   */
-  scoreTypeShortError?: boolean;
-}
-
-export enum CardSet {
-  fibonacci = 'Fibonacci',
-  powersOfTwo = 'Powers of 2',
-  customCardSet = 'Custom CardSet',
-}
-
-export const cardSetData = [
-  {
-    title: 'Fibonacci ( 0, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ?, Pass )',
-    value: CardSet.fibonacci,
-  },
-  {
-    title: 'Powers of 2 ( 0, 1, 2, 4, 8, 16, 32, 64, ?, Pass )',
-    value: CardSet.powersOfTwo,
-  },
-  {
-    title: 'Set your custom deck',
-    value: CardSet.customCardSet,
-  },
-];
 
 const useStyles = makeStyles((theme) => ({
   select: {
@@ -74,9 +44,9 @@ const GameSettingsSection = () => {
   const [errors, setErrors] = useState({} as IGameSettingsErrors);
 
   const validate = () => {
-    if (inputSettings.scoreType === '') {
+    if (!inputSettings.scoreType) {
       setErrors({ ...errors, scoreTypeError: true });
-    } else if (inputSettings.scoreTypeShort === '') {
+    } else if (!inputSettings.scoreTypeShort) {
       setErrors({ ...errors, scoreTypeShortError: true });
     } else {
       const newErrors = { ...errors };
